Move files to new test dir structure
